refactor: extract chunk_store_lock.c
Move graph-lock primitives, reload-from-disk, and lock/refresh APIs
into a dedicated TU. Promote helpers shared with open/commit
(csFileLock*, csReloadFromDisk*, csFileSizeByName,
csDiskStateMatchesMemory) via chunk_store_int.h. Wire build,
amalgamation, and lint (presence + ≤1200).
